pyproject.toml
[tool.poetry]
name = "counter_app"
version = "0.1.0"
description = ""
authors = ["DZhuribeda <zhurbikk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
uvicorn = "^0.17.6"
dependency-injector = "^4.38.0"
aioredis = "^2.0.1"
prometheus-fastapi-instrumentator = "^5.7.1"
structlog = "^21.5.0"
opentelemetry-distro = "^0.29b0"
opentelemetry-instrumentation-aws-lambda = "0.29b0"
opentelemetry-instrumentation-dbapi = "0.29b0"
opentelemetry-instrumentation-logging = "0.29b0"
opentelemetry-instrumentation-sqlite3 = "0.29b0"
opentelemetry-instrumentation-urllib = "0.29b0"
opentelemetry-instrumentation-wsgi = "0.29b0"
opentelemetry-instrumentation-asgi = "0.29b0"
opentelemetry-instrumentation-fastapi = "0.29b0"
opentelemetry-instrumentation-requests = "0.29b0"
opentelemetry-instrumentation-urllib3 = "0.29b0"
opentelemetry-exporter-otlp = "^1.10.0"
opentelemetry-propagator-b3 = "^1.10.0"
PyJWT = {version = "^2.4.0", extras = ["crypto"]}
cryptography = "36.0.2"
edgedb = "^0.23.0"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
black = {version = "^22.1.0", allow-prereleases = true}
requests = "^2.27.1"
pytest-asyncio = "^0.18.3"
mypy = "^0.960"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
namespace_packages = true

[tool.poetry.scripts]
server = "counter_app.asgi:server"